Use all XML formats for study import test.  Refactor study tests to allow for using import to setup most tests.

+++ api/src/org/labkey/api/security/UserManager.java
@@ -555,7 +555,7 @@ public class UserManager
                 scope.rollbackTransaction();
         }
 
-        clearUserList(userId);
+        clearUserList(userId.intValue());
 
         User user = getUser(userId.intValue());
         fireAddUser(user);
